+++ src/lib.rs
@@ -4,10 +4,10 @@ mod tests;
 mod wf_buffer;
 mod wf_codec;
 mod wf_core;
-mod wf_models;
+mod wf_json;
 
 pub fn encode_from_json<T: AsRef<str>>(json: T) -> Result<String, String> {
-    let message: wf_models::WhiteflagFieldValues =
+    let message: wf_json::WhiteflagFieldValues =
         serde_json::from_str(json.as_ref()).expect("deserialization error");
     Ok(wf_core::creator::encode(&message.fields))
 }
