+++ src/wf_json/test.rs
@@ -1,4 +1,4 @@
-use crate::{decode_from_hex, wf_models::deserialize::WhiteflagFieldValues};
+use crate::{decode_from_hex, wf_json::deserialize::WhiteflagFieldValues};
 use serde_json::json;
 
 #[test]
